Fixed vulnerabilities and a breaking change.
There were up to 127 vulnerabilities when the dependencies were
installed (with "npm install"). Eventually, the number came down
to zero by:

1. Updating/upgrading packages with "npm audit fix --force". The
only breaking change followed was an error caused by a lowercase
letter which started the React component's
"getSelectedLocatoinId" name. So, changed its occurences to
"GetSelectedLocatoinId".

2. Moving "react-scripts" package from "dependencies" to
"devDependencies" and overwriting the version of "@svgr/webpack"
to be "^6.5.1". More at:
https://github.com/facebook/create-react-app/issues/12132#issuecomment-1130249584
https://stackoverflow.com/questions/71282206/github-dependabot-alert-inefficient-regular-expression-complexity-in-nth-check

3. Updating the dependencies ("async" and "jquery") of
 "node_modules/@fmidev/metolib" (in "package-lock.json").